update socket.io example
加入用户认证，需登录后才可成功建立连接

+++ src/main/java/server/service/LoginService.java
@@ -5,5 +5,5 @@ import server.db.primary.dto.login.LoginDTO;
 public interface LoginService {
     LoginDTO selectUserByUsername(String username, String password);
 
-    LoginDTO selectUserById(Long id);
+    LoginDTO selectUserById(Integer id);
 }
